requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
librosa==0.10.2
numpy==1.26.4
soundfile==0.12.1
openai-whisper==20231117
boto3==1.34.69
httpx==0.27.0
